move detectorResult to interface
