+++ thirdeye-detectionpipeline/src/test/java/ai/startree/thirdeye/detectionpipeline/operator/PostProcessorOperatorTest.java
@@ -29,7 +29,8 @@ import ai.startree.thirdeye.detectionpipeline.DetectionRegistry;
 import ai.startree.thirdeye.detectionpipeline.OperatorContext;
 import ai.startree.thirdeye.detectionpipeline.PlanNodeContext;
 import ai.startree.thirdeye.detectionpipeline.PostProcessorRegistry;
-import ai.startree.thirdeye.detectionpipeline.operator.AnomalyDetectorOperatorResult.Builder;
+import ai.startree.thirdeye.spi.detection.v2.AnomalyDetectorOperatorResult;
+import ai.startree.thirdeye.spi.detection.v2.AnomalyDetectorOperatorResult.Builder;
 import ai.startree.thirdeye.spi.datalayer.TemplatableMap;
 import ai.startree.thirdeye.spi.datalayer.bao.DatasetConfigManager;
 import ai.startree.thirdeye.spi.datalayer.bao.EventManager;
